fix: remove trim suffix from resource_id in proxy
ref: https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/issues/747
For MSAL (v2.0 endpoint) asking an access token for a resource that accepts a v1.0 access token, Azure AD parses the desired audience
from the requested scope by taking everything before the last slash and using it as the resource identifier.
For example, if the scope is "https://vault.azure.net/.default", the resource identifier is "https://vault.azure.net".
If the scope is "http://database.windows.net//.default", the resource identifier is "http://database.windows.net/".

Signed-off-by: Anish Ramasekar <anish.ramasekar@gmail.com>