Support Microsoft Graph Endpoints for AzureStackCloud
When using the AzureStackCloud environment, the desired Microsoft Graph
endpoint/scope might not be one of the hard coded values for
PublicCloud, USGovernmentCloud, ChinaCloud, or GermanCloud. If the
msGraphEndpoint map doesn't contain an entry for the provided Azure
environment, use the ResourceIdentifier value for MicrosoftGraph from
the provided Azure environment. This allows users to override the
endpoint/scope when needed.