feat(helmchart): make the webhook container's securityContext configurable
In some environments, a hardcoded container securityContext can cause problems.
E.g. in OpenShift, a container's `securityContext.runAsUser` needs to be set to a namespace-specific
UID.

This commit moves the currently configured securityContext to the Helm values so
that the default securityContext stays the same, but can be overridden if needed.